pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nps-lens"
version = "0.1.0"
description = "NPS Lens — análisis avanzado de NPS térmico + VoC multi-fuente (MVP Senda MX)."
readme = "README.md"
requires-python = ">=3.9,<3.10"
license = {text = "Proprietary"}
authors = [{name = "NPS Lens"}]
dependencies = [
  "pandas>=2.0,<2.3",
  "numpy>=1.24,<2.0",
  "pydantic>=1.10,<2.0",
  "python-dotenv>=1.0,<2.0",
  "typer>=0.12,<0.13",
  "rich>=13.0,<14.0",
  "scikit-learn>=1.3,<1.6",
  "statsmodels>=0.14,<0.15",
  "networkx>=3.1,<3.4",
  "ruptures>=1.1.8,<2.0",
  "streamlit>=1.31,<1.40",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
  "pytest-cov>=5.0,<6.0",
  "mypy>=1.8,<1.12",
  "types-python-dateutil",
  "ruff>=0.3,<0.7",
  "black>=24.2,<25.0",
  "pre-commit>=3.6,<4.0",
]

[project.scripts]
nps-lens = "nps_lens.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E","F","I","B","UP","SIM","PL","RUF"]
ignore = ["PLR0913","PLR2004"]
fix = true

[tool.mypy]
python_version = "3.9"
strict = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["ruptures.*","networkx.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-q --cov=src/nps_lens --cov-report=term-missing --cov-fail-under=80"
testpaths = ["tests"]
